src/main/java/dev/deyve/algorithmsjava/utils/NumberUtils.java
package dev.deyve.algorithmsjava.utils;

import java.util.List;
import java.util.Objects;

public class NumberUtils {

    /**
     * Sum of Integers
     *
     * @param integers List of Integers
     * @return sum
     */
    static int sumOfIntegers(List<Integer> integers) {

        if (Objects.isNull(integers)) {
            throw new IllegalArgumentException("List cannot be null");
        }

        return integers
                .stream()
                .filter(Objects::nonNull)
                .mapToInt(Integer::intValue)
                .sum();
    }

}
